jsonrpc: add RPCServer.AliasMethod method
Adds this method and supporting struct field aliasedMethods
to provide feature support for method aliasing.

Methods are only queried by their alias when
the provided method name is not found in the
original registry, making the aliases essentially
fallback values.

If no alias is found for a given method,
the error 'method not found' is returned
in the same way as it was before this feature
was added.

Note that this method was added as a solution
to handling the implementation of rpc.discover
method at filecoin-project/lotus.

The RPC specification defines rpc.-prefixed
methods as 'system extensions', and handles
them as their own class of methods apart from
other API methods. As such, further development
may be interested in changing or extending
method aliasing to specification-congruent
handling of system extension methods.

Date: 2020-11-21 09:11:35-06:00
Signed-off-by: meows <b5c6@protonmail.com>